+++ src/App.tsx
@@ -2,6 +2,7 @@ import { useState } from 'react';
 import { Home } from './components/Home';
 import { ToolShell } from './components/ToolShell';
 import { NowPlayingBar } from './components/NowPlayingBar';
+import { EnergyButton } from './components/EnergyButton';
 import { getTool } from './tools/registry';
 
 function App() {
@@ -10,6 +11,7 @@ function App() {
 
   return (
     <>
+      <EnergyButton />
       {activeTool ? (
         <ToolShell
           icon={activeTool.meta.icon}