+++ src/main.tsx
@@ -4,11 +4,14 @@ import './amplifyConfig'
 import './index.css'
 import App from './App.tsx'
 import { WhiteNoiseProvider } from './context/WhiteNoiseContext'
+import { EnergyProvider } from './context/EnergyContext'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
-    <WhiteNoiseProvider>
-      <App />
-    </WhiteNoiseProvider>
+    <EnergyProvider>
+      <WhiteNoiseProvider>
+        <App />
+      </WhiteNoiseProvider>
+    </EnergyProvider>
   </StrictMode>,
 )